+++ testing/integration/src/common/tasks/stop.rs
