+++ testing/integration/src/tasks/block/submitter.rs
@@ -1,4 +1,4 @@
-use crate::common::{daemon::ClientManager, tasks::Task};
+use crate::{common::daemon::ClientManager, tasks::Task};
 use async_channel::Sender;
 use async_trait::async_trait;
 use kaspa_core::warn;
