+++ testing/integration/src/tasks/subscription/full.rs
@@ -1,5 +1,5 @@
-use crate::common::{
-    daemon::ClientManager,
+use crate::{
+    common::daemon::ClientManager,
     tasks::{
         notify::notification_drainer::NotificationDrainerTask,
         subscription::{
